pyproject.toml
[project]
name = "hirelatam-assessment"
version = "0.1.0"
description = "Launch & Fundraise Intelligence Dashboard - HireLatam 2026 assessment"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.50.0",
    "streamlit>=1.40.0",
    "pydantic>=2.9.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.8.0",
]

[tool.uv]
package = false

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests", "evals"]
addopts = "-q"
